move FileSystemProjectTree to separate file
