+++ src/python/pants/base/project_tree.py
@@ -8,9 +8,7 @@ from __future__ import (absolute_import, division, generators, nested_scopes, pr
 import logging
 import os
 from abc import abstractmethod
-from glob import glob1
 
-from pants.util.dirutil import safe_walk
 from pants.util.meta import AbstractClass
 
 
@@ -61,31 +59,3 @@ class ProjectTree(AbstractClass):
 
   def __hash__(self):
     return hash(self.__repr__())
-
-
-class FileSystemProjectTree(ProjectTree):
-  def glob1(self, dir_relpath, glob):
-    return glob1(os.path.join(self.build_root, dir_relpath), glob)
-
-  def content(self, file_relpath):
-    with open(os.path.join(self.build_root, file_relpath), 'rb') as source:
-      return source.read()
-
-  def isdir(self, relpath):
-    return os.path.isdir(os.path.join(self.build_root, relpath))
-
-  def isfile(self, relpath):
-    return os.path.isfile(os.path.join(self.build_root, relpath))
-
-  def exists(self, relpath):
-    return os.path.exists(os.path.join(self.build_root, relpath))
-
-  def walk(self, relpath, topdown=True):
-    for root, dirs, files in safe_walk(os.path.join(self.build_root, relpath), topdown=topdown):
-      yield os.path.relpath(root, self.build_root), dirs, files
-
-  def __eq__(self, other):
-    return other and (type(other) == type(self)) and (self.build_root == other.build_root)
-
-  def __repr__(self):
-    return '{}({})'.format(self.__class__.__name__, self.build_root)
