+++ tests/python/pants_test/build_graph/test_build_configuration.py
@@ -10,7 +10,7 @@ import unittest
 from contextlib import contextmanager
 
 from pants.base.build_file import BuildFile
-from pants.base.project_tree import FileSystemProjectTree
+from pants.base.file_system_project_tree import FileSystemProjectTree
 from pants.build_graph.build_configuration import BuildConfiguration
 from pants.build_graph.build_file_aliases import BuildFileAliases, TargetMacro
 from pants.build_graph.target import Target
